Extract S2S flow control and timing reports from handler.py
handler.py mixed RPC handling with two self-contained concerns:

- S2SLinkTelemetry + AdaptivePushConcurrency (sliding-window link stats
  and the self-tuning push concurrency limiter) move to s2s_flow.py
- the ~330-line run-end [TIMING_SUMMARY]/[PAPER_TIMING_TABLE]/[PIPELINE_*]
  report generator moves to timing_summary.py as a pure function; the
  handler keeps a thin wrapper that pops its session dicts

No behavior change; handler.py shrinks from 3098 to 2596 lines.
